Make location option fully answer option / Fix JS bugs / Fix multiplication
Location options had their own template which was integrated to answer option since location options are using answer option models aswell. Due to this change there had to be some JS changes which also fixed some bugs, e.g. a bug for customizeSearch where a map controller wasn't found when trying to fire this method. clickListener for select locations became redundant because of these changes which allowed a removal of an extra function. Regarding answer options there was a bug in the JS -file that didn't disable / delete the location option fields when question type was changed, so if geojson and tooltip direction had values it caused multiplication of the answer option elements when changing the question type back to an answer option question type.