Fix map resets for map question types
When you would have all the default values for map_locations on the default placement, it would create a marker in the center of the map. Now if default values are there, the map is empty with the default position. Select_locations was also using a wrong method to center the map if geojson was empty, it has now been changed correctly